+++ config/prod.exs
@@ -21,6 +21,7 @@ config :phoenix, serve_endpoints: true
 
 # Do not print debug messages in production
 config :logger, :console, level: :warn
+config :logger, :ex_syslogger, level: :warn
 
 # ## SSL Support
 #